(PC-29252) refactor(share): Rename files
