+++ src/features/share/helpers/shareAppModalInformation.tsx
